+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/OperationSupport.java
@@ -16,7 +16,6 @@
 package io.fabric8.kubernetes.client.dsl.internal;
 
 import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import io.fabric8.kubernetes.api.model.DeleteOptions;
 import io.fabric8.kubernetes.api.model.DeletionPropagation;
 import io.fabric8.kubernetes.api.model.HasMetadata;
@@ -37,8 +36,9 @@ import io.fabric8.kubernetes.client.dsl.base.PatchType;
 import io.fabric8.kubernetes.client.http.HttpClient;
 import io.fabric8.kubernetes.client.http.HttpRequest;
 import io.fabric8.kubernetes.client.http.HttpResponse;
+import io.fabric8.kubernetes.client.impl.BaseClient;
 import io.fabric8.kubernetes.client.utils.KubernetesResourceUtil;
-import io.fabric8.kubernetes.client.utils.Serialization;
+import io.fabric8.kubernetes.client.utils.KubernetesSerialization;
 import io.fabric8.kubernetes.client.utils.URLUtils;
 import io.fabric8.kubernetes.client.utils.Utils;
 import org.slf4j.Logger;
@@ -46,7 +46,6 @@ import org.slf4j.LoggerFactory;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.InterruptedIOException;
 import java.lang.reflect.Type;
 import java.net.HttpURLConnection;
@@ -55,7 +54,6 @@ import java.net.URL;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
@@ -69,7 +67,6 @@ public class OperationSupport {
   public static final String STRATEGIC_MERGE_JSON_PATCH = "application/strategic-merge-patch+json";
   public static final String JSON_MERGE_PATCH = "application/merge-patch+json";
 
-  protected static final ObjectMapper JSON_MAPPER = Serialization.jsonMapper();
   private static final Logger LOG = LoggerFactory.getLogger(OperationSupport.class);
   private static final String CLIENT_STATUS_FLAG = "CLIENT_STATUS_FLAG";
 
@@ -253,7 +250,7 @@ public class OperationSupport {
     String itemNs = KubernetesResourceUtil.getNamespace((HasMetadata) item);
 
     if (Utils.isNotNullOrEmpty(namespace) && Utils.isNotNullOrEmpty(itemNs) && !namespace.equals(itemNs)) {
-      item = Serialization.clone(item);
+      item = getKubernetesSerialization().clone(item);
       KubernetesResourceUtil.setNamespace((HasMetadata) item, namespace);
     }
     return item;
@@ -318,7 +315,7 @@ public class OperationSupport {
     }
 
     HttpRequest.Builder requestBuilder = httpClient.newHttpRequestBuilder()
-        .delete(JSON, JSON_MAPPER.writeValueAsString(deleteOptions)).url(requestUrl);
+        .delete(JSON, getKubernetesSerialization().asJson(deleteOptions)).url(requestUrl);
 
     return handleResponse(requestBuilder, KubernetesResource.class);
   }
@@ -338,7 +335,7 @@ public class OperationSupport {
   protected <T, I> T handleCreate(I resource, Class<T> outputType) throws InterruptedException, IOException {
     resource = correctNamespace(resource);
     HttpRequest.Builder requestBuilder = httpClient.newHttpRequestBuilder()
-        .post(JSON, JSON_MAPPER.writeValueAsString(resource))
+        .post(JSON, getKubernetesSerialization().asJson(resource))
         .url(getResourceURLForWriteOperation(getResourceUrl(checkNamespace(resource), null)));
     return handleResponse(requestBuilder, outputType);
   }
@@ -356,7 +353,7 @@ public class OperationSupport {
   protected <T> T handleUpdate(T updated, Class<T> type) throws IOException {
     updated = correctNamespace(updated);
     HttpRequest.Builder requestBuilder = httpClient.newHttpRequestBuilder()
-        .put(JSON, JSON_MAPPER.writeValueAsString(updated))
+        .put(JSON, getKubernetesSerialization().asJson(updated))
         .url(getResourceURLForWriteOperation(getResourceUrl(checkNamespace(updated), checkName(updated))));
     return handleResponse(requestBuilder, type);
   }
@@ -389,7 +386,7 @@ public class OperationSupport {
         }
       }
       // we can't omit status unless this is not a status operation and we know this has a status subresource
-      patchForUpdate = PatchUtils.jsonDiff(current, updated, false);
+      patchForUpdate = PatchUtils.jsonDiff(current, updated, false, getKubernetesSerialization());
       if (patchContext == null) {
         patchContext = new PatchContext.Builder().withPatchType(PatchType.JSON).build();
       }
@@ -405,7 +402,7 @@ public class OperationSupport {
           }
         }
       }
-      patchForUpdate = Serialization.asJson(updated);
+      patchForUpdate = getKubernetesSerialization().asJson(updated);
       current = updated; // use the updated to determine the path
     }
     return handlePatch(patchContext, current, patchForUpdate, type);
@@ -445,7 +442,7 @@ public class OperationSupport {
   protected <T> T handleScale(String resourceUrl, T scale, Class<T> scaleType) throws InterruptedException, IOException {
     HttpRequest.Builder requestBuilder = httpClient.newHttpRequestBuilder().uri(resourceUrl + "/scale");
     if (scale != null) {
-      requestBuilder.put(JSON, JSON_MAPPER.writeValueAsString(scale));
+      requestBuilder.put(JSON, getKubernetesSerialization().asJson(scale));
     }
     return handleResponse(requestBuilder, scaleType);
   }
@@ -462,7 +459,7 @@ public class OperationSupport {
   protected Status handleDeploymentRollback(String resourceUrl, DeploymentRollback deploymentRollback)
       throws InterruptedException, IOException {
     HttpRequest.Builder requestBuilder = httpClient.newHttpRequestBuilder().uri(resourceUrl + "/rollback").post(JSON,
-        JSON_MAPPER.writeValueAsString(deploymentRollback));
+        getKubernetesSerialization().asJson(deploymentRollback));
     return handleResponse(requestBuilder, Status.class);
   }
 
@@ -474,10 +471,9 @@ public class OperationSupport {
    * @param <T> template argument provided
    *
    * @return returns a deserialized object as api server response of provided type.
-   * @throws InterruptedException Interrupted Exception
    * @throws IOException IOException
    */
-  protected <T> T handleGet(URL resourceUrl, Class<T> type) throws InterruptedException, IOException {
+  protected <T> T handleGet(URL resourceUrl, Class<T> type) throws IOException {
     HttpRequest.Builder requestBuilder = httpClient.newHttpRequestBuilder().url(resourceUrl);
     return handleResponse(requestBuilder, type);
   }
@@ -563,7 +559,7 @@ public class OperationSupport {
       try {
         assertResponseCode(request, response);
         if (type != null && type.getType() != null) {
-          return Serialization.unmarshal(new ByteArrayInputStream(response.body()), type);
+          return getKubernetesSerialization().unmarshal(new ByteArrayInputStream(response.body()), type);
         } else {
           return null;
         }
@@ -598,9 +594,10 @@ public class OperationSupport {
     String customMessage = config.getErrorMessages().get(statusCode);
 
     if (customMessage != null) {
-      throw requestFailure(request, createStatus(statusCode, combineMessages(customMessage, createStatus(response))));
+      throw requestFailure(request,
+          createStatus(statusCode, combineMessages(customMessage, createStatus(response, getKubernetesSerialization()))));
     } else {
-      throw requestFailure(request, createStatus(response));
+      throw requestFailure(request, createStatus(response, getKubernetesSerialization()));
     }
   }
 
@@ -614,7 +611,7 @@ public class OperationSupport {
     return customMessage;
   }
 
-  public static Status createStatus(HttpResponse<?> response) {
+  public static Status createStatus(HttpResponse<?> response, KubernetesSerialization kubernetesSerialization) {
     String statusMessage = "";
     int statusCode = response != null ? response.code() : 0;
     if (response == null) {
@@ -623,7 +620,7 @@ public class OperationSupport {
       try {
         String bodyString = response.bodyString();
         if (Utils.isNotNullOrEmpty(bodyString)) {
-          Status status = JSON_MAPPER.readValue(bodyString, Status.class);
+          Status status = kubernetesSerialization.unmarshal(bodyString, Status.class);
           if (status != null) {
             if (status.getCode() == null) {
               status = new StatusBuilder(status).withCode(statusCode).build();
@@ -631,7 +628,7 @@ public class OperationSupport {
             return status;
           }
         }
-      } catch (IOException e) {
+      } catch (IOException | KubernetesClientException | IllegalArgumentException e) {
         // ignored
       }
       if (response.message() != null) {
@@ -691,23 +688,6 @@ public class OperationSupport {
     return requestException(request, e, null);
   }
 
-  protected static <T> T unmarshal(InputStream is) {
-    return Serialization.unmarshal(is);
-  }
-
-  protected static <T> T unmarshal(InputStream is, final Class<T> type) {
-    return Serialization.unmarshal(is, type);
-  }
-
-  protected static <T> T unmarshal(InputStream is, TypeReference<T> type) {
-    return Serialization.unmarshal(is, type);
-  }
-
-  protected static <T> Map<String, Object> getObjectValueAsMap(T object) {
-    return JSON_MAPPER.convertValue(object, new TypeReference<Map<String, Object>>() {
-    });
-  }
-
   public Config getConfig() {
     return config;
   }
@@ -763,7 +743,7 @@ public class OperationSupport {
       if (payload instanceof String) {
         body = (String) payload;
       } else if (payload != null) {
-        body = Serialization.asJson(payload);
+        body = getKubernetesSerialization().asJson(payload);
       }
       HttpRequest request = withRequestTimeout(httpClient.newHttpRequestBuilder().uri(uri).method(method, JSON, body)).build();
       HttpResponse<R1> response = waitForResult(httpClient.sendAsync(request, result));
@@ -772,6 +752,10 @@ public class OperationSupport {
     } catch (IOException e) {
       throw KubernetesClientException.launderThrowable(e);
     }
+  }
+
+  public KubernetesSerialization getKubernetesSerialization() {
+    return context.getClient().adapt(BaseClient.class).getKubernetesSerialization();
   }
 
 }