+++ openshift-client-api/src/main/java/io/fabric8/openshift/client/dsl/TemplateResource.java
@@ -15,85 +15,87 @@
  */
 package io.fabric8.openshift.client.dsl;
 
+import io.fabric8.kubernetes.api.model.KubernetesList;
 import io.fabric8.kubernetes.client.dsl.Resource;
+import io.fabric8.openshift.api.model.Template;
 import io.fabric8.openshift.client.ParameterValue;
 
 import java.io.File;
 import java.io.InputStream;
 import java.util.Map;
 
-public interface TemplateResource<T, L> extends Resource<T> {
+public interface TemplateResource extends Resource<Template> {
 
   /**
    * Process a template reading parameter values from a {@link File} in json or yml format.
-   * 
+   *
    * @param f The specified {@link File}.
    * @return
    */
-  L process(File f);
+  KubernetesList process(File f);
 
   /**
    * Process a template reading parameter values from an {@link InputStream} in json or yml format.
-   * 
+   *
    * @param is The specified {@link InputStream}.
    * @return
    */
-  L process(InputStream is);
+  KubernetesList process(InputStream is);
 
   /**
    * Process a template reading parameter values from am {@link Map}.
-   * 
+   *
    * @param map The specified {@link Map}.
    * @return
    */
-  L process(Map<String, String> map);
+  KubernetesList process(Map<String, String> map);
 
   /**
    * Process a template with the specified {@link ParameterValue}s.
-   * 
+   *
    * @param values The specified {@link ParameterValue}s.
    * @return
    */
-  L process(ParameterValue... values);
+  KubernetesList process(ParameterValue... values);
 
   /**
    * Process a template locally reading parameter values from a {@link File} in json or yml format.
    * This kind of processing is performed locally, without communicating with the server (e.g for generating values using
    * expressions).
-   * 
+   *
    * @param f The specified {@link File}.
    * @return
    */
-  L processLocally(File f);
+  KubernetesList processLocally(File f);
 
   /**
    * Process a template locally reading parameter values from an {@link InputStream} in json or yml format.
    * This kind of processing is performed locally, without communicating with the server (e.g for generating values using
    * expressions).
-   * 
+   *
    * @param is The specified {@link InputStream}.
    * @return
    */
-  L processLocally(InputStream is);
+  KubernetesList processLocally(InputStream is);
 
   /**
    * Process a template locally reading parameter values from am {@link Map}.
    * This kind of processing is performed locally, without communicating with the server (e.g for generating values using
    * expressions).
-   * 
+   *
    * @param map The specified {@link Map}.
    * @return
    */
-  L processLocally(Map<String, String> map);
+  KubernetesList processLocally(Map<String, String> map);
 
   /**
    * Process a template with the specified {@link ParameterValue}s.
    * This kind of processing is performed locally, without communicating with the server (e.g for generating values using
    * expressions).
-   * 
+   *
    * @param values The specified {@link ParameterValue}s.
    * @return
    */
-  L processLocally(ParameterValue... values);
+  KubernetesList processLocally(ParameterValue... values);
 
 }